refactor: move PageContent components to their own folder
