+++ packages/website/components/PageContent/index.ts
@@ -1,0 +1,2 @@
+export { PageContent } from './PageContent';
+export type { HeadingType } from './TableOfContent';
